+++ src/components/navbar/customNavbar/mobileNavbar.tsx
@@ -1,30 +1,49 @@
 import { Link, useNavigate } from "react-router-dom";
 import Avatar from "@/assets/no-profile 1.png";
 import { Bell, ChevronLeft, Search } from "lucide-react";
-import { ReactNode } from "react";
+import { FormEvent, ReactNode, useState } from "react";
 
 interface Props {
   setOpen?: (value : boolean) => void
   open? : boolean
   children? : ReactNode
+  onSearch : (e : FormEvent<HTMLFormElement>) => void
 }
 
-export const MobileNavbar = ({setOpen , open}:Props) => {
-  
+export const MobileNavbar = ({setOpen , open, onSearch}:Props) => {
+  const [openSearch, setOpenSearch] = useState(false)
+  const navigate = useNavigate()
   return (
     <>
       <div className="flex justify-between p-1">
-        <button onClick={() => setOpen!(!open)}>
-          {/* <CircleUser /> */}
-          <img src={Avatar} width={35} className="rounded-full" alt="" />
-        </button>
-        <Link to={"/home"} className="font-bold text-xl my-auto text-blue-500">
-          HoYoLAB
-        </Link>
-        <div className="flex justify-between gap-5">
-          <Search className="my-auto w-5 h-5 aspect-square" />
-          <Bell className="my-auto w-5 h-5 aspect-square"/>
-        </div>
+        {
+          openSearch ? 
+          <>
+            <form action="" className="w-full mx-1" onSubmit={onSearch}>
+              <input className="w-full text-xs py-1 px-3 font-semibold rounded-full bg-slate-700 border border-transparent focus:border-blue-500 focus:bg-primary focus:outline-none hover:border-blue-500 hover:bg-primary" placeholder="Telusuri konten yang menarik" type="text" name="search"/>
+            </form>
+            <button className="text-blue-600 text-xs font-semibold" onClick={() => {
+              navigate('/')
+              setOpenSearch(false)
+            }}>Batalkan</button>
+          </>
+          :
+          <>
+            <button onClick={() => setOpen!(!open)}>
+              {/* <CircleUser /> */}
+              <img src={Avatar} width={35} className="rounded-full" alt="" />
+            </button>
+            <Link to={"/home"} className="font-bold text-xl my-auto text-blue-500">
+              HoYoLAB
+            </Link>
+            <div className="flex justify-between gap-5">
+              <button className="my-auto" onClick={() => setOpenSearch(true)}>
+                <Search className="w-5 h-5 aspect-square" />
+              </button>
+              <Bell className="my-auto w-5 h-5 aspect-square"/>
+            </div>
+          </>
+        }
       </div>
       
     </>
